+++ ConcordiumWallet/Features/TransferTokenFlow/Models/CIS2TokenTransferModel.swift
@@ -10,6 +10,7 @@ import Foundation
 import Combine
 import BigInt
 import UIKit
+import Concordium
 
 enum TokenTransferNotifyDestination {
     case none
@@ -34,7 +35,7 @@ final class CIS2TokenTransferModel {
     let notifyDestination: TokenTransferNotifyDestination
     
     private var cancellables = [AnyCancellable]()
-    private let dependencyProvider: AccountsFlowCoordinatorDependencyProvider
+    private let dependencyProvider: ServicesProvider
     private let passwordDelegate: RequestPasswordDelegate
     private var onTxSuccess: (String) -> Void
     private var onTxReject: () -> Void
@@ -46,7 +47,7 @@ final class CIS2TokenTransferModel {
     init(
         tokenType: CXTokenType,
         account: AccountDataType,
-        dependencyProvider: AccountsFlowCoordinatorDependencyProvider,
+        dependencyProvider: ServicesProvider,
         notifyDestination: TokenTransferNotifyDestination,
         passwordDelegate: RequestPasswordDelegate = DummyRequestPasswordDelegate(),
         memo: Memo?,
@@ -84,6 +85,8 @@ final class CIS2TokenTransferModel {
                     .map { balance -> BigDecimal in
                         return .init(BigInt(stringLiteral: balance.balance), cis2Token.metadata.decimals ?? 0)
                     } ?? .zero(cis2Token.metadata.decimals ?? 0)
+        case .plt(let token):
+            return .init(BigInt(stringLiteral: token.token.tokenState.totalSupply.value), token.token.tokenState.totalSupply.decimals)
         }
     }
     
@@ -94,6 +97,8 @@ final class CIS2TokenTransferModel {
                 return try await self.getCIS2TxCost(cIS2Token, amount: amountTokenSend)
             case .ccd:
                 return try await getCCDTxCost()
+        case .plt(let token):
+            return try await getCCDTxCost()
         }
     }
     
@@ -108,6 +113,8 @@ final class CIS2TokenTransferModel {
                     await self.updateCIS2TransferConst(cIS2Token, amount: amount)
                 case .ccd:
                     await updateCCDTransferCost()
+            case .plt(let token):
+                await updateCCDTransferCost()
             }
             await self.updateMaxAmount()
         }).store(in: &cancellables)
@@ -140,23 +147,28 @@ final class CIS2TokenTransferModel {
                     self.tokenGeneralBalance = .zero
                     return
                 }
-                    self.maxAmountTokenSend = .init(BigInt(stringLiteral: balance.balance), token.metadata.decimals ?? 0)
+            self.maxAmountTokenSend = .init(BigInt(stringLiteral: balance.balance), token.metadata.decimals ?? 0)
                     self.tokenGeneralBalance = .init(BigInt(stringLiteral: balance.balance), token.metadata.decimals ?? 0)
+        case .plt(let token):
+            await updateCCDTransferCost()
+                
+            self.maxAmountTokenSend = .init(BigInt(stringLiteral: token.tokenAccountState.balance.value), token.tokenAccountState.balance.decimals)
+            self.tokenGeneralBalance = .init(BigInt(stringLiteral: token.tokenAccountState.balance.value), token.tokenAccountState.balance.decimals)
         }
     }
     
-    func executeTransaction() async throws -> AnyPublisher<TransferEntity, Error> {
-        return try await callTransaction().tryMap { [weak self] entity in
-            guard let self = self else { return entity }
-            switch self.notifyDestination {
-                case .legacyQrConnect:
-//                    self.legacyQRConnectService?.sendPaymentMessage(hash: entity.submissionId ?? "")
-                    self.onTxSuccess(entity.submissionId ?? "")
-                case .none: break
-            }
-            return entity
-        }.eraseToAnyPublisher()
-    }
+//    func executeTransaction() async throws -> AnyPublisher<TransferEntity, Error> {
+//        return try await callTransaction().tryMap { [weak self] entity in
+//            guard let self = self else { return entity }
+//            switch self.notifyDestination {
+//                case .legacyQrConnect:
+////                    self.legacyQRConnectService?.sendPaymentMessage(hash: entity.submissionId ?? "")
+//                    self.onTxSuccess(entity.submissionId ?? "")
+//                case .none: break
+//            }
+//            return entity
+//        }.eraseToAnyPublisher()
+//    }
 }
 
 extension CIS2TokenTransferModel {
@@ -196,60 +208,140 @@ extension CIS2TokenTransferModel {
     }
 }
 
+//extension CIS2TokenTransferModel {
+//    private func callTransaction() async throws -> AnyPublisher<TransferEntity, Error> {
+//        switch self.tokenType {
+//            case .cis2(let cIS2Token):
+//                guard let recipient = self.recipient,
+//                      let transaferCost = self.transaferCost
+//                else { return .fail(TransferTokenError.insuficientData) }
+//                return try await transferCis2Token(cIS2Token, amount: self.amountTokenSend, to: recipient, txCost: transaferCost)
+//            case .ccd:
+//                return try await simpleTransferCCDToken()
+//        case .plt(let token):
+//  
+//            return try await transferPLT(token)
+//        }
+//    }
+//    
+//    private func transferPLT(_ token: AccountPLTToken) async throws -> AnyPublisher<TransferEntity, Error> {
+//        dependencyProvider.concordiumClient().transferPLT(
+//            token: token,
+//            sender: <#T##AccountAddress#>,
+//            amount: <#T##String#>,
+//            decimals: <#T##UInt16#>,
+//            receiver: <#T##AccountAddress#>,
+//            keys: <#T##AccountKeys#>,
+//            memo: <#T##Memo?#>
+//        )
+//    }
+//    
+//    @MainActor
+//    private func simpleTransferCCDToken() async throws -> AnyPublisher<TransferEntity, Error> {
+//        var transfer = TransferDataTypeFactory.create()
+//        transfer.transferType = .simpleTransfer
+//        transfer.amount = String(self.amountTokenSend.value)
+//        transfer.fromAddress = self.account.address
+//        transfer.toAddress = self.recipient ?? ""
+//        transfer.cost = self.transaferCost?.cost ?? "100000"
+//        transfer.energy = self.transaferCost?.energy ?? 0
+//        transfer.memo = self.memo?.data.hexDescription
+//
+//        return dependencyProvider.transactionsService()
+//            .performTransfer(transfer, from: self.account, requestPasswordDelegate: self.passwordDelegate)
+//            .tryMap { transferDataType -> TransferEntity in
+//                _ = try self.dependencyProvider.storageManager().storeTransfer(transferDataType)
+//                return transferDataType as! TransferEntity
+//            }
+//            .eraseToAnyPublisher()
+//    }
+//    
+//    @MainActor
+//    private func transferCis2Token(_ token: CIS2Token, amount: BigDecimal, to: String, txCost: TransferCost) async throws -> AnyPublisher<TransferEntity, Error> {
+//        let serializedTransferParams = try MobileWalletFacade().serializeTokenTransferParameters(input: TokenTransferParameters(tokenId: token.tokenId, amount: String(self.amountTokenSend.value), from: self.account.address, to: to))
+//        
+//        var transfer = TransferDataTypeFactory.create()
+//        transfer.transferType = .transferUpdate
+//        transfer.from = self.account.address
+//        transfer.toAddress = to
+//        transfer.expiry = Date().addingTimeInterval(10 * 60)
+//        transfer.energy = txCost.energy
+//        transfer.memo = self.memo?.data.hexDescription
+//        
+//        transfer.receiveName = token.contractName + ".transfer"
+//        transfer.params = serializedTransferParams
+//
+//        return dependencyProvider.transactionsService()
+//            .performTransferUpdate(transfer, from: self.account, contractAddress: .init(index: token.contractAddress.index, subindex: token.contractAddress.subindex), requestPasswordDelegate: self.passwordDelegate)
+//            .tryMap { transferDataType -> TransferEntity in
+//                _ = try self.dependencyProvider.storageManager().storeTransfer(transferDataType)
+//                return transferDataType as! TransferEntity
+//            }
+//            .eraseToAnyPublisher()
+//    }
+//}
+
 extension CIS2TokenTransferModel {
-    private func callTransaction() async throws -> AnyPublisher<TransferEntity, Error> {
-        switch self.tokenType {
+    func executeTransfer() async throws {
+            guard let recipient = self.recipient,
+                  let transaferCost = self.transaferCost
+            else { throw TransferTokenError.insuficientData }
+
+            switch self.tokenType {
             case .cis2(let cIS2Token):
-                guard let recipient = self.recipient,
-                      let transaferCost = self.transaferCost
-                else { return .fail(TransferTokenError.insuficientData) }
-                return try await transferCis2Token(cIS2Token, amount: self.amountTokenSend, to: recipient, txCost: transaferCost)
+                try await ececuteTransferCIS2(token: cIS2Token, recipient: recipient)
             case .ccd:
-                return try await simpleTransferCCDToken()
+                try await executeTransferCCD(recipient: recipient)
+            case .plt(_):
+                break
+            }
         }
-    }
-    
-    @MainActor
-    private func simpleTransferCCDToken() async throws -> AnyPublisher<TransferEntity, Error> {
-        var transfer = TransferDataTypeFactory.create()
-        transfer.transferType = .simpleTransfer
-        transfer.amount = String(self.amountTokenSend.value)
-        transfer.fromAddress = self.account.address
-        transfer.toAddress = self.recipient ?? ""
-        transfer.cost = self.transaferCost?.cost ?? "100000"
-        transfer.energy = self.transaferCost?.energy ?? 0
-        transfer.memo = self.memo?.data.hexDescription
 
-        return dependencyProvider.transactionsService()
-            .performTransfer(transfer, from: self.account, requestPasswordDelegate: self.passwordDelegate)
-            .tryMap { transferDataType -> TransferEntity in
-                _ = try self.dependencyProvider.storageManager().storeTransfer(transferDataType)
-                return transferDataType as! TransferEntity
-            }
-            .eraseToAnyPublisher()
-    }
-    
-    @MainActor
-    private func transferCis2Token(_ token: CIS2Token, amount: BigDecimal, to: String, txCost: TransferCost) async throws -> AnyPublisher<TransferEntity, Error> {
-        let serializedTransferParams = try MobileWalletFacade().serializeTokenTransferParameters(input: TokenTransferParameters(tokenId: token.tokenId, amount: String(self.amountTokenSend.value), from: self.account.address, to: to))
-        
-        var transfer = TransferDataTypeFactory.create()
-        transfer.transferType = .transferUpdate
-        transfer.from = self.account.address
-        transfer.toAddress = to
-        transfer.expiry = Date().addingTimeInterval(10 * 60)
-        transfer.energy = txCost.energy
-        transfer.memo = self.memo?.data.hexDescription
-        
-        transfer.receiveName = token.contractName + ".transfer"
-        transfer.params = serializedTransferParams
+        @MainActor
+        func executeTransferCCD(recipient: String) async throws {
+            let pwHash = try await self.passwordDelegate.requestUserPassword(keychain: dependencyProvider.keychainWrapper())
+            guard
+                let encryptedAccountDataKey = account.encryptedAccountData,
+                let accountKeys = try? dependencyProvider.storageManager().getPrivateAccountKeys(key: encryptedAccountDataKey, pwHash: pwHash).get()
+            else { throw WalletError.invalidInput }
 
-        return dependencyProvider.transactionsService()
-            .performTransferUpdate(transfer, from: self.account, contractAddress: .init(index: token.contractAddress.index, subindex: token.contractAddress.subindex), requestPasswordDelegate: self.passwordDelegate)
-            .tryMap { transferDataType -> TransferEntity in
-                _ = try self.dependencyProvider.storageManager().storeTransfer(transferDataType)
-                return transferDataType as! TransferEntity
+            var concordiumMemo: Concordium.Memo? = nil
+            
+            if let memoData = memo?.data {
+                concordiumMemo = Concordium.Memo(memoData)
             }
-            .eraseToAnyPublisher()
-    }
+            
+            let submittedTransaction = try await dependencyProvider.concordiumClient().transferCCD(
+                sender: AccountAddress(base58Check: self.account.address),
+                amount: CCD.init(microCCD: MicroCCDAmount(Double(self.amountTokenSend.value))),
+                receiver: AccountAddress(base58Check: recipient),
+                keys: accountKeys,
+                memo: concordiumMemo
+            )
+            // TODO: - should handle tx status
+            let transactionStatus = try await dependencyProvider.concordiumClient().getTransactionStatus(submittedTransaction.hash)
+            print(submittedTransaction)
+            print(transactionStatus)
+        }
+
+        @MainActor
+        func ececuteTransferCIS2(token: CIS2Token, recipient: String) async throws {
+            let pwHash = try await self.passwordDelegate.requestUserPassword(keychain: dependencyProvider.keychainWrapper())
+            guard
+                let encryptedAccountDataKey = account.encryptedAccountData,
+                let accountKeys = try? dependencyProvider.storageManager().getPrivateAccountKeys(key: encryptedAccountDataKey, pwHash: pwHash).get()
+            else { throw WalletError.invalidInput }
+
+            let submittedTransaction = try await dependencyProvider.concordiumClient().transferCIS2(
+                sender: AccountAddress(base58Check: self.account.address),
+                receiver: AccountAddress(base58Check: recipient),
+                keys: accountKeys,
+                contractAddress: Concordium.ContractAddress(index: UInt64(token.contractAddress.index), subindex: UInt64(token.contractAddress.subindex)),
+                tokenId: token.tokenId,
+                amount: self.amountTokenSend.value
+            )
+            let transactionStatus = try await dependencyProvider.concordiumClient().getTransactionStatus(submittedTransaction.hash)
+            print(submittedTransaction)
+            print(transactionStatus)
+        }
 }